+++ packages/ui/src/components/Alert/index.ts
@@ -1,4 +1,4 @@
 export { default as H0Alert } from './H0Alert.vue'
 export { default as H0AlertDialog } from './H0AlertDialog.vue'
-export type { H0AlertDialogBackdrop, H0AlertDialogTone, H0AlertTone } from './Alert.types'
+export type { H0AlertDialogBackdrop, H0AlertDialogTone, H0AlertTone, H0AlertVariant } from './Alert.types'
 export type { H0AlertDialogEmits, H0AlertDialogProps, H0AlertEmits, H0AlertProps } from './Alert.types'