+++ packages/ui/src/components/ContentState/index.ts
@@ -1,0 +1,2 @@
+export { default as H0ContentState } from './H0ContentState.vue'
+export type * from './ContentState.types'